Update test.py
I was trying to run the code and it couldn't run... Then I thought "what's the point of testing a gan? is it to see it's accuracy? no, it's not a CNN, it must only be about downloading the trained the model (tf.train.Supervisor(logdir="gan/train_logs/ .....)) and see what images could be generated with the trained generator, so I should maybe save the images in another file to see them (the logwriter in gan/test_logs/). 
Now it works and it also displays the images with the step being displayed, same as the modification in train.py file.
Thanks for this wonderful code!